Rename solid to feature mod
