+++ packages/cadmium/src/feature/point.rs
